metomi/rose#731: make the splash screen usage nicer
