+++ lib/python/rose/gtk/util.py
@@ -511,125 +511,6 @@ class TooltipTreeView(gtk.TreeView):
         return self.get_tooltip(view, row_iter, col_index, tip)
 
 
-class SplashScreen(gtk.Window):
-
-    """Run a splash screen that receives update information."""
-    
-    BACKGROUND_COLOUR = "white"  # Same as logo background.
-    PADDING = 10
-    SUB_PADDING = 5
-    FONT_DESC = "8"
-    PULSE_FRACTION = 0.05
-    TIME_WAIT_FINISH = 500  # Milliseconds.
-    TIME_IDLE_BEFORE_PULSE = 3000  # Milliseconds.
-    TIME_INTERVAL_PULSE = 50  # Milliseconds.
-
-    def __init__(self, logo_path, title, total_number_of_events):
-        super(SplashScreen, self).__init__()
-        self.set_title(title)
-        self.set_decorated(False)
-        self.stopped = False
-        try:
-            locator = rose.resource.ResourceLocator(paths=sys.path)
-            icon_path = locator.locate('etc/images/rose-icon-trim.png')
-        except rose.resource.ResourceError:
-            pass
-        else:
-            self.set_icon_from_file(icon_path)
-        self.modify_bg(gtk.STATE_NORMAL,
-                       gtk.gdk.color_parse(self.BACKGROUND_COLOUR))
-        self.set_gravity(gtk.gdk.GRAVITY_CENTER)
-        self.set_position(gtk.WIN_POS_CENTER)
-        main_vbox = gtk.VBox()
-        main_vbox.show()
-        image = gtk.image_new_from_file(logo_path)
-        image.show()
-        image_hbox = gtk.HBox()
-        image_hbox.show()
-        image_hbox.pack_start(image, expand=False, fill=True)
-        main_vbox.pack_start(image_hbox, expand=False, fill=True)
-        self._is_progress_bar_pulsing = False
-        self._progress_fraction = 0.0
-        self.progress_bar = gtk.ProgressBar()
-        self.progress_bar.set_pulse_step(self.PULSE_FRACTION)
-        self.progress_bar.show()
-        self.progress_bar.modify_font(pango.FontDescription(self.FONT_DESC))
-        self.progress_bar.set_ellipsize(pango.ELLIPSIZE_END)
-        self._progress_message = None
-        self.event_count = 0.0
-        self.total_number_of_events = float(total_number_of_events)
-        progress_hbox = gtk.HBox(spacing=self.SUB_PADDING)
-        progress_hbox.show()
-        progress_hbox.pack_start(self.progress_bar, expand=True, fill=True,
-                                 padding=self.SUB_PADDING)
-        main_vbox.pack_start(progress_hbox, expand=False, fill=False,
-                             padding=self.PADDING)
-        self.add(main_vbox)
-        if self.total_number_of_events > 0:
-            self.show()
-        while gtk.events_pending():
-            gtk.main_iteration()
-
-    def update(self, event, no_progress=False, new_total_events=None):
-        """Show text corresponding to an event."""
-        text = str(event)
-        if new_total_events is not None:
-            self.total_number_of_events = new_total_events
-            self.event_count = 0.0
-            
-        if not no_progress:
-            self.event_count += 1.0
-            
-        if self.total_number_of_events == 0:
-            fraction = 1.0
-        else:
-            fraction = min([1.0, self.event_count /
-                                 self.total_number_of_events])
-        self._stop_pulse()
-        
-        if not no_progress:
-            gobject.idle_add(self.progress_bar.set_fraction, fraction)
-            self._progress_fraction = fraction
-            
-        self.progress_bar.set_text(text)
-        self._progress_message = text
-        gobject.timeout_add(self.TIME_IDLE_BEFORE_PULSE,
-                            self._start_pulse, fraction, text)
-
-        if fraction == 1.0 and not no_progress:
-            gobject.timeout_add(self.TIME_WAIT_FINISH,
-                                lambda: self.finish())
-
-        while gtk.events_pending():
-            gtk.main_iteration()
-
-    def _start_pulse(self, idle_fraction, idle_message):
-        """Start the progress bar pulsing (moving side-to-side)."""
-        if (self._progress_message != idle_message or
-            self._progress_fraction != idle_fraction):
-            return False
-        self._is_progress_bar_pulsing = True
-        gobject.timeout_add(self.TIME_INTERVAL_PULSE,
-                            self._pulse)
-        return False
-
-    def _stop_pulse(self):
-        self._is_progress_bar_pulsing = False
-    
-    def _pulse(self):
-        if self._is_progress_bar_pulsing:
-            self.progress_bar.pulse()
-        while gtk.events_pending():
-            gtk.main_iteration()
-        return self._is_progress_bar_pulsing
-
-    def finish(self):
-        """Delete the splash screen."""
-        self.stopped = True
-        gobject.idle_add(self.destroy)
-        return False
-
-
 class DialogProcess(object):
 
     """Run a forked process and display a dialog while it runs.
